+++ example/lib/pages/flutter_logo.dart
@@ -24,8 +24,6 @@ class FlutterLogoPage extends StatelessWidget {
 }
 
 class FlutterLogoPainter extends InteractivePainter {
-  final Rough rough = Rough();
-
   Paint stroke = Paint()
     ..strokeWidth = 2
     ..isAntiAlias = true
@@ -43,12 +41,11 @@ class FlutterLogoPainter extends InteractivePainter {
 
   @override
   void paintRough(canvas, size) {
-    FillerConfig fillerConfig = FillerConfig(
+    FillerConfig fillerConfig = FillerConfig.build(
       hachureAngle: 90,
       drawConfig: drawConfig,
       fillWeight: 10,
       hachureGap: 5,
-//      zigzagOffset: 10,
     );
 
     Generator gen = Generator(drawConfig, HachureFiller(fillerConfig));
@@ -62,11 +59,11 @@ class FlutterLogoPainter extends InteractivePainter {
 
     canvas
       ..translate(translateX, translateY)
-      ..scale(scale);
-    rough.draw(canvas, gen.polygon([PointD(37, 128), PointD(9, 101), PointD(100, 10), PointD(156, 10)]), stroke, fillPaint);
-    canvas.translate(-4, -4);
-    rough.draw(canvas, gen.polygon([PointD(156, 94), PointD(100, 94), PointD(50, 141), PointD(79, 170)]), stroke, fillPaint);
-    canvas.translate(6, 6);
-    rough.draw(canvas, gen.polygon([PointD(79, 170), PointD(100, 191), PointD(156, 191), PointD(107, 142)]), stroke, fillPaint);
+      ..scale(scale)
+      ..drawRough(gen.polygon([PointD(37, 128), PointD(9, 101), PointD(100, 10), PointD(156, 10)]), stroke, fillPaint)
+      ..translate(-4, -4)
+      ..drawRough(gen.polygon([PointD(156, 94), PointD(100, 94), PointD(50, 141), PointD(79, 170)]), stroke, fillPaint)
+      ..translate(6, 6)
+      ..drawRough(gen.polygon([PointD(79, 170), PointD(100, 191), PointD(156, 191), PointD(107, 142)]), stroke, fillPaint);
   }
 }