+++ example/lib/pages/interactive_arc.dart
@@ -41,9 +41,9 @@ class ArcPainter extends InteractivePainter {
 
   @override
   void paintRough(Canvas canvas, Size size) {
-    Generator generator = Generator(drawConfig, HachureFiller(const FillerConfig().copyWith(drawConfig: drawConfig)));
+    Generator generator = Generator(drawConfig, HachureFiller(FillerConfig.build().copyWith(drawConfig: drawConfig)));
     double s = min(size.width, size.height);
     Drawable figure = generator.arc(size.width / 2, size.height / 2, s * 0.8, s * 0.8, pi * 0.2, pi * 1.8, true);
-    Rough().draw(canvas, figure, pathPaint, fillPaint);
+    canvas.drawRough(figure, pathPaint, fillPaint);
   }
 }