+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.1] - 2020-May-5.
+
+* `FillConfig no has a builder`
+* `Rough.draw` is now `drawRough`, a extension method of ``Draw
+* Basic external api documentation
+
 ## [0.1.0+2] - 2020-May-5.
 
 * Basic documentation